backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==11.0.0
opencv-python==4.10.0.84
scikit-image==0.24.0
numpy==2.1.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.0